+++ src/test/java/com/bitbank/controllers/BankAccountTransactionsControllerTest.java
@@ -9,9 +9,12 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import java.math.BigDecimal;
+import java.util.stream.Stream;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
@@ -64,7 +67,7 @@ class BankAccountTransactionsControllerTest {
 	@WithMockUser("username")
 	@Test
 	void testCanIndex() throws Exception {
-		
+
 		@SuppressWarnings("unchecked")
 		Page<BankAccountTransactionsEntity> items = mock(Page.class);
 		when(bankAccountTransactionsService.index(0, 1, "id", "desc")).thenReturn(items);
@@ -84,6 +87,28 @@ class BankAccountTransactionsControllerTest {
 				.content(objectMapper.writeValueAsBytes(transactionToSave))).andExpect(status().isCreated())
 				.andExpect(jsonPath("$.id", is((int) id)))
 				.andExpect(jsonPath("$.amount", is(transactionToSave.getAmount())));
+	}
+
+	@WithMockUser("username")
+	@ParameterizedTest
+	@MethodSource("getInvalidBankAccountTransactions")
+	void testCanCatchException(BankAccountTransactionsEntity invalidBankAccountTransactionsEntity) throws Exception {
+
+		mvc.perform(post("/api/v1/bank-account-transactions/").contentType(MediaType.APPLICATION_JSON)
+				.content(objectMapper.writeValueAsBytes(invalidBankAccountTransactionsEntity)))
+				.andExpect(status().isBadRequest());
+
+	}
+	
+	private static Stream<BankAccountTransactionsEntity> getInvalidBankAccountTransactions() {
+		// Create a valid entity
+		var validBankAccountTransactionsEntity = validBankAccountTransactionsEntity(new BigDecimal(100), "ok");
+		
+		return Stream.of(
+				validBankAccountTransactionsEntity.toBuilder().amount(null).build(),
+				validBankAccountTransactionsEntity.toBuilder().purpose(null).build(),
+				validBankAccountTransactionsEntity.toBuilder().purpose("").build()
+				);
 	}
 
 	private static BankAccountTransactionsEntity validBankAccountTransactionsEntity(Long id, BigDecimal amount,