+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 + Remove unused imports
 + Fix "Type safety: The expression of type Page needs unchecked conversion to conform to Page<BankAccountTransactionsEntity>"
 + Return JSON if not authorized
++ Add Bank Account validation
 
 ## 0.10.0
 + Change swagger server